update spatie/laravel permission, seed

+++ app/Modules/Core/Routes/permission.php
@@ -3,13 +3,13 @@
 use Illuminate\Support\Facades\Route;
 use App\Modules\Core\PermissionController;
 
-Route::get('/export', [PermissionController::class, 'export']);
-Route::post('/import', [PermissionController::class, 'import']);
-Route::post('/bulk-delete', [PermissionController::class, 'bulkDestroy']);
-Route::get('/stats', [PermissionController::class, 'stats']);
-Route::get('/', [PermissionController::class, 'index']);
-Route::get('/{permission}', [PermissionController::class, 'show']);
-Route::post('/', [PermissionController::class, 'store']);
-Route::put('/{permission}', [PermissionController::class, 'update']);
-Route::patch('/{permission}', [PermissionController::class, 'update']);
-Route::delete('/{permission}', [PermissionController::class, 'destroy']);
+Route::get('/export', [PermissionController::class, 'export'])->middleware('permission:permissions.export,web');
+Route::post('/import', [PermissionController::class, 'import'])->middleware('permission:permissions.import,web');
+Route::post('/bulk-delete', [PermissionController::class, 'bulkDestroy'])->middleware('permission:permissions.bulkDestroy,web');
+Route::get('/stats', [PermissionController::class, 'stats'])->middleware('permission:permissions.stats,web');
+Route::get('/', [PermissionController::class, 'index'])->middleware('permission:permissions.index,web');
+Route::get('/{permission}', [PermissionController::class, 'show'])->middleware('permission:permissions.show,web');
+Route::post('/', [PermissionController::class, 'store'])->middleware('permission:permissions.store,web');
+Route::put('/{permission}', [PermissionController::class, 'update'])->middleware('permission:permissions.update,web');
+Route::patch('/{permission}', [PermissionController::class, 'update'])->middleware('permission:permissions.update,web');
+Route::delete('/{permission}', [PermissionController::class, 'destroy'])->middleware('permission:permissions.destroy,web');
